code/backend/requirements.txt
# Backend dependencies
flask
sentence_transformers
python-dotenv
psycopg2
sqlalchemy